changes made for now

+++ src/Model/Job.java
@@ -5,11 +5,11 @@ public class Job {
     private String title;
     private String description;
     private String location;
-    private float salary;
+    private double salary;
     private String postDate;
     private int companyId;
     
-    public Job(int id, String title, String description, String location, float salary, String postDate, int companyId) {
+    public Job(int id, String title, String description, String location, double salary, String postDate, int companyId) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -24,7 +24,7 @@ public class Job {
     public String getTitle() { return title; }
     public String getDescription() { return description; }
     public String getLocation() { return location; }
-    public float getSalary() { return salary; }
+    public double getSalary() { return salary; }
     public String getPostDate() { return postDate; }
     public int getCompanyId() { return companyId; }
     
@@ -33,7 +33,7 @@ public class Job {
     public void setTitle(String title) { this.title = title; }
     public void setDescription(String description) { this.description = description; }
     public void setLocation(String location) { this.location = location; }
-    public void setSalary(float salary) { this.salary = salary; }
+    public void setSalary(double salary) { this.salary = salary; }
     public void setPostDate(String postDate) { this.postDate = postDate; }
     public void setCompanyId(int companyId) { this.companyId = companyId; }
 }
